requirements.txt
annotated-types==0.6.0
anthropic==0.23.1
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.3
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
iniconfig==2.0.0
ipython==8.18.1
jedi==0.19.1
jsonpath-ng==1.6.1
matplotlib-inline==0.1.6
outcome==1.3.0.post0
packaging==24.0
parso==0.8.3
pexpect==4.9.0
playwright==1.42.0
pluggy==1.4.0
ply==3.11
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
pyee==11.0.1
Pygments==2.17.2
PySocks==1.7.1
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
text-unidecode==1.3
tokenizers==0.15.2
tomli==2.0.1
tqdm==4.66.2
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
